Add usage information
Sometimes, people see a nicely named repository which seems to align with their specific needs. In some cases, these potential users are put off by the lack of usage information. Not everyone is experimentally inclined, and therefore one might not try using some standard tools to explore the inner workings of the tool. This commit aims to fix that problem, providing passers-by with the initial step-by-step description of how to use the current state of the software, as displayed in this code base.